Modify the kbdd to run as a daemon on Windows within the user account.  It will then use GetLastInputInfo to check keyboard input every five seconds or so and send an update to the startd if it actually detects a change in the input.  The idle checking code in the startd has also been modified to not bother using GetLastInputInfo, though we may need to change this so that a personal Condor on Windows will still be able to keep track of idle time without the help of kbdd.  Still need to modify installer script to set condor_kbdd to be run at user login.  Ticket #714.

+++ src/condor_sysapi/last_x_event.c
@@ -27,6 +27,5 @@ sysapi_last_xevent(void)
 {
 	sysapi_internal_reconfig();
 
-	/* XXX why isn't _sysapi_last_x_event a time_t? */
-	_sysapi_last_x_event = (int)time(NULL);
+	_sysapi_last_x_event = time(NULL);
 }
